Issue #SB-23803 fix: refactored code
